# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for i-Manage application.",
        "title": "i-Manage API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/ishop": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Create a new iShop (store card)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ishop"
                ],
                "summary": "Create iShop",
                "parameters": [
                    {
                        "description": "iShop creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateIShopRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.IShopProfile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ishop/{id}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Get iShop details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ishop"
                ],
                "summary": "Get iShop",
                "parameters": [
                    {
                        "type": "string",
                        "description": "iShop ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.IShopProfile"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Update iShop profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ishop"
                ],
                "summary": "Update iShop",
                "parameters": [
                    {
                        "type": "string",
                        "description": "iShop ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateIShopRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Delete iShop and remove from all iComs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ishop"
                ],
                "summary": "Delete iShop",
                "parameters": [
                    {
                        "type": "string",
                        "description": "iShop ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ishop/{id}/memberships": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "List all iCom memberships for a shop",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ishop"
                ],
                "summary": "List iShop Memberships",
                "parameters": [
                    {
                        "type": "string",
                        "description": "iShop ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.IShopMembershipInfo"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.CreateIShopRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "banner": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "imageUrls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "industry": {
                    "description": "Industry",
                    "type": "string"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "description": "Contact",
                    "type": "string"
                },
                "province": {
                    "description": "Address",
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "subIndustry": {
                    "type": "string"
                },
                "ward": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
...
        "models.IShopMembershipInfo": {
            "type": "object",
            "properties": {
                "icomId": {
                    "type": "string"
                },
                "icomLogo": {
                    "type": "string"
                },
                "icomName": {
                    "type": "string"
                },
                "joinedDate": {
                    "type": "string"
                },
                "rank": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "score": {
                    "description": "Membership score/timestamp",
                    "type": "number"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.IShopProfile": {
            "type": "object",
            "properties": {
                "banner": {
                    "type": "string"
                },
                "cardType": {
                    "type": "string"
                },
                "created": {
                    "description": "Audit fields",
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "icoms": {
                    "description": "RediSearch indexing",
                    "type": "string"
                },
                "id": {
                    "description": "System fields",
                    "type": "string"
                },
                "imageUrls": {
                    "description": "JSON array string",
                    "type": "string"
                },
                "industry": {
                    "description": "Industry fields",
                    "type": "string"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "logo": {
                    "type": "string"
                },
                "modified": {
                    "type": "string"
                },
                "name": {
                    "description": "Profile fields",
                    "type": "string"
                },
                "packageId": {
                    "type": "string"
                },
                "phone": {
                    "description": "Contact fields",
                    "type": "string"
                },
                "private": {
                    "type": "string"
                },
                "province": {
                    "description": "Address fields",
                    "type": "string"
                },
                "status": {
                    "description": "Configuration",
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "subIndustry": {
                    "type": "string"
                },
                "ward": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateIShopRequest": {
            "type": "object",
            "properties": {
                "banner": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "imageUrls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "industry": {
                    "description": "Industry",
                    "type": "string"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "description": "Contact",
                    "type": "string"
                },
                "province": {
                    "description": "Address",
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "subIndustry": {
                    "type": "string"
                },
                "ward": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "token",
            "in": "cookie"
        }
    }
}